+++ NEWS.md
@@ -32,7 +32,9 @@
   e.g. `median` that sometimes returns `integer` sometimes `numeric`. (#893). 
 
 * update hybrid evaluation to handle `$` better (#1134).   
-  
+
+* Fixed `bind_rows` `POSIXct` bug (#1125). 
+
 # dplyr 0.4.2
 
 This is a minor release containing fixes for a number of crashes and issues identified by R CMD CHECK. There is one new "feature": dplyr no longer complains about unrecognised attributes, and instead just copies them over to the output.